fix: use toolchain file while building dependencies
This commit make the cargo build command that builds the dependencies
actually adhere to the toolchain file if it was present in the recipe.

Before this it was ignored and thus the next call to cargo build would
rebuild all dependencies again, but then with what is actually contained
in the toolchain file. This was because the environment variable
'RUSTUP_TOOLCHAIN' was already set and so the toolchain file was not
considered, see here for details:
https://rust-lang.github.io/rustup/overrides.html