feat: Add --ignore-manifest cook arg
One way to speed up builds involving large monomorphized dependencies is
to separate their usage into a separate crate, which is then cached.

It's helpful if that can be a package in the same workspace, but AFAICT
cargo-chef doesn't currently have a mechanism to extend its caching to
some workspace members.

So to support that, let's add an `--ignore-manifest` flag to request
that cargo-chef doesn't overwrite that `Cargo.toml` file or clean up the
build artifacts from that package.